Add unittest and coverage infr. for Python and R
Along with vscode extension recommendations, complementary OS-indepedent vscode tasks, necessary gitignore/Rbuildignore changes and so on.

There is a bit more to say here, since adapting everything with the vscode extension tooling so that the whole structure works seamlessly was not trivial. Just not to miss one particular magic (aka bugfix): the cobertura xml outputter of covr package of R do not properly add the sources tag information for the file root location. To remedy that, I added a sed replace that should work without any issues. See tasks.json for more details.

Some magic was also necessary to work with Rscript --so many quotation issues. Yet everything should be resolved with this commit.