Adjust outlined tip thicknesses
Bar alignment and edge thickness values are now dependent on the scores. To allow configuring the edge thicknesses in Python, inner and outer patches are also separated as in R.

The main issue with the patches in Python is that the winding matters, and the inner/outer paths need to be wound inversely to create holes in the outer patches. Removing the clipping hack also allowed to omit setting edge thicknesses to zero for outlined=False cases.

Also the default thickness value is fixed in the documentation.